+++ CTPUSHTEMPLATES/build.gradle
@@ -1,11 +1,11 @@
 plugins {
-    alias(libs.plugins.android.library)
-    alias(libs.plugins.jetbrains.kotlin.android)
+    id 'com.android.library'
+    id 'kotlin-android'
 }
 
 android {
-    namespace 'com.karthik.ctpushtemplates'
-    compileSdk 34
+    namespace 'com.clevertap.ctcustomtemplates'
+    compileSdk 33
 
     defaultConfig {
         minSdk 24
