+++ CTPUSHTEMPLATES/src/androidTest/java/com/karthik/ctpushtemplates/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.karthik.ctpushtemplates
+package com.clevertap.ctpushtemplates
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.karthik.ctpushtemplates.test", appContext.packageName)
+        assertEquals("com.clevertap.ctpushtemplates.test", appContext.packageName)
     }
 }
