+++ CTPUSHTEMPLATES/src/main/java/com/clevertap/ctcustomtemplates/pn/PushNotificationListener.java
@@ -1,4 +1,4 @@
-package com.karthik.ctpushtemplates.pn;
+package com.clevertap.ctcustomtemplates.pn;
 
 public interface PushNotificationListener {
     void onPushRendered();
