+++ CTPUSHTEMPLATES/src/test/java/com/karthik/ctpushtemplates/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.karthik.ctpushtemplates
+package com.clevertap.ctcustomtemplates
 
 import org.junit.Test
