+++ gradle.properties
@@ -1,3 +1,4 @@
+android.nonFinalResIds=false
 # Project-wide Gradle settings.
 # IDE (e.g. Android Studio) users:
 # Gradle settings configured through the IDE *will override*
